+++ libsa4py/cst_extractor.py
@@ -1,7 +1,8 @@
 from libsa4py.cst_visitor import Visitor
-from libsa4py.representations import ModuleInfo
+from libsa4py.representations import ModuleInfo, create_output_seq
 from libsa4py.cst_transformers import TypeAdder, SpaceAdder, StringRemover, CommentAndDocStringRemover, NumberRemover,\
     TypeAnnotationRemover
+from libsa4py.nl_preprocessing import normalize_module_code
 from libsa4py.exceptions import ParseError
 
 import libcst as cst
@@ -14,8 +15,7 @@ class Extractor:
 
     @staticmethod
     def extract(program: str,
-                program_types: cst.metadata.type_inference_provider.PyreData = None) -> dict:
-
+                program_types: cst.metadata.type_inference_provider.PyreData = None) -> ModuleInfo:
         try:
             parsed_program = cst.parse_module(program)
         except Exception as e:
@@ -50,4 +50,5 @@ class Extractor:
         v_typed = v_typed.visit(v_space)
 
         return ModuleInfo(v.imports, v.module_variables, v.module_variables_use, v.cls_list, v.fns,
-                          v_untyped.code, v_typed.code).to_dict()
+                          normalize_module_code(v_untyped.code), create_output_seq(normalize_module_code(v_typed.code)),
+                          v.module_type_annot_cove)