+++ libsa4py/cst_pipeline.py
@@ -156,8 +156,8 @@ class Pipeline:
                                                       filename) if self.use_pyre else None
 
                     project_analyzed_files[project_id]["src_files"][f_relative] = \
-                        self.apply_nlp_transf(Extractor().extract(read_file(filename), pyre_data_file)) if self.nlp_transf \
-                            else Extractor.extract(read_file(filename), pyre_data_file)
+                        self.apply_nlp_transf(Extractor().extract(read_file(filename), pyre_data_file).to_dict()) if self.nlp_transf \
+                            else Extractor.extract(read_file(filename), pyre_data_file).to_dict()
 
                     project_analyzed_files[project_id]["src_files"][f_relative]['set'] = f_split