+++ libsa4py/cst_transformers.py
@@ -44,6 +44,9 @@ class NumberRemover(cst.CSTTransformer):
     def leave_Integer(self, original_node: cst.Integer, updated_node: cst.Integer):
         return cst.SimpleString(value="\"[number]\"")
 
+    def leave_Imaginary(self, original_node: cst.Imaginary, updated_node: cst.Imaginary):
+        return cst.SimpleString(value="\"[number]\"")
+
 
 class TypeAnnotationRemover(cst.CSTTransformer):
     """
@@ -524,6 +527,11 @@ class SpaceAdder(cst.CSTTransformer):
         return updated_node.with_changes(
             whitespace_after=cst.SimpleWhitespace(' '),
             whitespace_before=cst.SimpleWhitespace(' ')
+        )
+
+    def leave_And(self, original_node, updated_node):
+        return updated_node.with_changes(
+            whitespace_after=cst.SimpleWhitespace(' ')
         )
 
     def leave_Plus(self, node, updated_node):