+++ tests/examples/imports.py
@@ -1,0 +1,13 @@
+"""
+This file contains various forms import statements for testing
+"""
+
+import math
+import typing as t
+from os.path import join, split
+from string import *
+
+
+def inner_imports():
+    import collections
+    collections.Counter()