+++ tests/examples/num_removal.py
@@ -1,0 +1,11 @@
+"""
+An example for removing numeric literals
+"""
+
+x = 12
+y = 3.14
+s = 1.3e+4
+c = 2+3j
+b = 0b100
+o = 0o320
+h = 0x12b