+++ tests/exp_outputs/removed_num.py
@@ -1,0 +1,11 @@
+"""
+An example for removing numeric literals
+"""
+
+x = "[number]"
+y = "[number]"
+s = "[number]"
+c = "[number]"+"[number]"
+b = "[number]"
+o = "[number]"
+h = "[number]"